+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## Unreleased
+## Ormolu 0.8.1.1
 
 * Add missing braces for case expressions in single‑line do blocks. [Issue
   1180](https://github.com/tweag/ormolu/issues/1180).
+
 * Fix the import grouping logic in the presence of imports with explicit
   levels. [Issue 1192](https://github.com/tweag/ormolu/issues/1192).